// src/app/models/constants/FieldStyleTypes.ts
export enum FieldTypes {
    placeholderText = "placeholderText",
    width = "width",
    height = "height",
    required = "required",
    borderStyle = "borderStyle",
    borderWidth = "borderWidth",
    fontSize = "fontSize",
    fontWeight = "fontWeight",
    borderColor = "borderColor",
    color = "textColor",
    backgroundColor = "backgroundColor",
    borderControl = "borderControl",
    labelText = "labelText",
    checkboxStyle = "checkboxStyle"
}